Add missing checks for NULL pointers
