+++ src/unix/apple/utils.rs
@@ -35,32 +35,6 @@ impl<T> Drop for CFReleaser<T> {
 unsafe impl<T> Send for CFReleaser<T> {}
 unsafe impl<T> Sync for CFReleaser<T> {}
 
-pub(crate) fn cstr_to_rust(c: *const c_char) -> Option<String> {
-    cstr_to_rust_with_size(c, None)
-}
-
-pub(crate) fn cstr_to_rust_with_size(c: *const c_char, size: Option<usize>) -> Option<String> {
-    if c.is_null() {
-        return None;
-    }
-    let mut s = match size {
-        Some(len) => Vec::with_capacity(len),
-        None => Vec::new(),
-    };
-    let mut i = 0;
-    unsafe {
-        loop {
-            let value = *c.offset(i) as u8;
-            if value == 0 {
-                break;
-            }
-            s.push(value);
-            i += 1;
-        }
-        String::from_utf8(s).ok()
-    }
-}
-
 pub(crate) fn vec_to_rust(buf: Vec<i8>) -> Option<String> {
     String::from_utf8(
         buf.into_iter()
