+++ src/unix/mod.rs
@@ -26,3 +26,4 @@ cfg_if::cfg_if! {
 
 pub(crate) mod network_helper;
 pub(crate) mod users;
+pub(crate) mod utils;
